+++ Micycle/Micycle/MiSchoolMenu.cs
@@ -9,9 +9,7 @@ namespace Micycle
 {
     class MiSchoolMenu : MiBuildingMenu
     {
-        public MiSchoolMenu(Micycle game, int center_x, int center_y)
-            : base(game, center_x, center_y)
-        {
-        }
+        public MiSchoolMenu(Micycle game, float center_x, float center_y)
+            : base(game, center_x, center_y) { }
     }
 }
